Install Jest testing framework dependency
Jest is a testing framework for testing JavaScript. Alongside some
libraries speific to testing React components, Jest could be called an
industry standard.

With Jest I would like to put up some example tests that can be used
for teaching so that we can eventually get testing coverage on the
project.